+++ architecture-templates/app/src/main/java/com/sfedu/recipebook/ui/recipe/RecipeScreen.kt
@@ -121,10 +121,10 @@ fun RecipeScreen(
                                             AlertDialog(
                                                 onDismissRequest = { openDialog.value = false},
                                                 title = { Text(stringResource(id = R.string.delete_recipe_alert)) },
-                                                confirmButton = {
+                                                dismissButton  = {
                                                     Button(
                                                         onClick = { openDialog.value = false },
-                                                        modifier = Modifier.width(116.dp),
+                                                        modifier = Modifier.width(126.dp),
                                                         shape = RoundedCornerShape(15.dp),
                                                         colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFE0F2F1), contentColor = Color.Black),
                                                         border = BorderStroke(2.dp, Color(0xFF4DB6AC))
@@ -132,12 +132,12 @@ fun RecipeScreen(
                                                         Text(stringResource(id = R.string.cancel_button))
                                                     }
                                                 },
-                                                dismissButton = {
+                                                confirmButton = {
                                                     Button(
                                                         onClick = { viewModel.deleteRecipe(recipe)
                                                             onNavigateToMain(); openDialog.value = false },
                                                         //openDialog.value = false ,
-                                                        modifier = Modifier.width(116.dp),
+                                                        modifier = Modifier.width(126.dp),
                                                         shape = RoundedCornerShape(15.dp),
                                                         colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFE0F2F1), contentColor = Color.Black),
                                                         border = BorderStroke(2.dp, Color(0xFF4DB6AC))
@@ -171,7 +171,7 @@ fun RecipeScreen(
                         Button(
                             onClick = { openDialog.value = true },
                             modifier = Modifier
-                                .width(126.dp)
+                                .width(140.dp)
                                 .padding(start = 5.dp),
                             shape = RoundedCornerShape(15.dp),
                             colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFE0F2F1), contentColor = Color.Black),
@@ -183,13 +183,13 @@ fun RecipeScreen(
                             AlertDialog(
                                 onDismissRequest = { openDialog.value = false},
                                 title = { Text(stringResource(id = R.string.reset_ingredients_alert)) },
-                                dismissButton = {
+                                confirmButton  = {
                                     Button(
                                         onClick = {
                                                 resetViewableIngredients(ingredients = ingredients)
                                                 onNavigateToRecipeView(); openDialog.value = false
                                         },
-                                        modifier = Modifier.width(116.dp),
+                                        modifier = Modifier.width(126.dp),
                                         shape = RoundedCornerShape(15.dp),
                                         colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFE0F2F1), contentColor = Color.Black),
                                         border = BorderStroke(2.dp, Color(0xFF4DB6AC))
@@ -197,10 +197,10 @@ fun RecipeScreen(
                                         Text(stringResource(id = R.string.reset_ingredients_button))
                                     }
                                 },
-                                confirmButton = {
+                                dismissButton = {
                                     Button(
                                         onClick = { openDialog.value = false },
-                                        modifier = Modifier.width(116.dp),
+                                        modifier = Modifier.width(126.dp),
                                         shape = RoundedCornerShape(15.dp),
                                         colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFE0F2F1), contentColor = Color.Black),
                                         border = BorderStroke(2.dp, Color(0xFF4DB6AC))
@@ -223,7 +223,7 @@ fun RecipeScreen(
                             onNavigateToCalc() },
 
                             modifier = Modifier
-                                .width(126.dp)
+                                .width(140.dp)
                                 .padding(end = 5.dp),
                             shape = RoundedCornerShape(15.dp),
                             colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFE0F2F1), contentColor = Color.Black),
